Crate fil_actor_interface

Modules

account

convert

cron

datacap

evm

init

market

miner

multisig

power

reward

system

verifreg

Macros

list_miners_for_state

parse_pending_transactions

This macro iterates over each transaction, decodes the transaction key using variable-length integer encoding, and constructs a Transaction object with the decoded data.

parse_pending_transactions_v3

This macro iterates over each transaction, decodes the transaction key, and constructs a Transaction object with additional processing for address and token formats using from_address_v3_to_v2 and from_token_v3_to_v2.

parse_pending_transactions_v4

This macro iterates over each transaction, assumes that transaction id’s are directly available and not encoded. It constructs Transaction objects with transformations for address and token data from version 4 to version 2 using from_address_v4_to_v2 and from_token_v4_to_v2.

Structs

AwardBlockRewardParams

FilterEstimate

Policy

Constants

BURNT_FUNDS_ACTOR_ADDR

CHAOS_ACTOR_ADDR

EPOCH_DURATION_SECONDS

Duration of each tipset epoch.

RESERVE_ADDRESS