Crate fil_actor_interface

Re-exports

pub use known_cids::KNOWN_CIDS;

pub use known_cids::KNOWN_CIDS;

Modules

account

cron

datacap

ethaccount

evm

init

known_cids

market

miner

multisig

paych

placeholder

power

reward

system

verifreg

Macros

cids_filename

list_miners_for_state

parse_pending_transactions

This macro iterates over each transaction, decodes the transaction key using variable-length integer encoding, and constructs a Transaction object with the decoded data.

parse_pending_transactions_v3

This macro iterates over each transaction, decodes the transaction key, and constructs a Transaction object with additional processing for address and token formats using from_address_v3_to_v2 and from_token_v3_to_v2.

parse_pending_transactions_v4

This macro iterates over each transaction, assumes that transaction id’s are directly available and not encoded. It constructs Transaction objects with transformations for address and token data from version 4 to version 2 using from_address_v4_to_v2 and from_token_v4_to_v2.

Structs

ActorCids

AwardBlockRewardParams

FilterEstimate

KNOWN_CIDS

NetworkManifest

Policy

Constants

BURNT_FUNDS_ACTOR_ADDR

CHAOS_ACTOR_ADDR

EPOCH_DURATION_SECONDS

Duration of each tipset epoch.

RESERVE_ADDRESS

Functions

is_account_actor

Returns true if the code belongs to an account actor.

is_eth_account_actor

Returns true if the code belongs to a ethereum account actor.

is_evm_actor

Returns true if the code belongs to an evm actor.

is_miner_actor

Returns true if the code belongs to a storage miner actor.

is_multisig_actor

Returns true if the code belongs to a multisig account actor.

is_paych_actor

Returns true if the code belongs to a payment channel actor.

is_placeholder_actor

Returns true if the code belongs to a placeholder actor.