Crate fil_actor_interface

Re-exports

pub use known_cids::KNOWN_CIDS;

pub use known_cids::KNOWN_CIDS;

Modules

account

cron

datacap

ethaccount

evm

init

known_cids

market

miner

multisig

placeholder

power

reward

system

verifreg

Structs

AwardBlockRewardParams

FilterEstimate

KNOWN_CIDS

NETWORK_MANIFESTS

NetworkManifest

Policy

Constants

BURNT_FUNDS_ACTOR_ADDR

CHAOS_ACTOR_ADDR

EPOCH_DURATION_SECONDS

Duration of each tipset epoch.

RESERVE_ADDRESS

Functions

is_account_actor

Returns true if the code belongs to an account actor.

is_eth_account_actor

Returns true if the code belongs to a ethereum account actor.

is_placeholder_actor

Returns true if the code belongs to a placeholder actor.